call method each and first
call each and first in #save method at variable item_id

+++ module-4/session-3/app/models/customer.rb
@@ -8,27 +8,45 @@ class Customer
   end
 
   def save
+    return false unless valid?
+
     client = create_db_client
     client.query("
         insert into customers (name, phone) values ('#{@name}', '#{@phone}');
-      ")
+      ".gsub(/\s+/, ' '))
+
+    true
   end
 
   def update
+    return false unless valid?
+
     client = create_db_client
     client.query("
         update customers
         set name='#{@name}', phone='#{@phone}'
         where id='#{@id}';
-      ")
+      ".gsub(/\s+/, ' '))
+
+    true
   end
 
   def delete
+    return false unless valid?
+
     client = create_db_client
     client.query("
         delete from customers
         where id='#{@id}';
-      ")
+      ".gsub(/\s+/, ' '))
+
+    true
+  end
+
+  def valid?
+    return false if @name.nil? || @phone.nil?
+
+    true
   end
 
   def self.all
@@ -36,15 +54,14 @@ class Customer
     rawData = client.query("
         select * from customers
       ")
-    
+
     convert_sql_to_array(rawData)
   end
 
   def self.convert_sql_to_array(rawData)
-    customers = Array.new
+    customers = []
     rawData.each do |data|
       data.transform_keys!(&:to_sym)
-      p data
       customer = Customer.new(data)
       customers << customer
     end
